Refactor one node down tests
